Add attribute buffers and more
This adds buffers to the attributes that allows actors to soak damage
before penalizing the attribute values.

With this change we also remodel the character sheet to fit everything
in.  As a result we also add an experience section to the main tab. The
critical injuries are moved to the main tab and are replaced with a new
actions bar on the combat tab.